rename BestDatabaseSequenceMapping -> SIFTSMapping
